# Optimize array pattern checktype