@nobu 2020-04-04
I saw 1.03x at maximum from the master.