@k-tsj 2020-04-04
Sorry for late response.

I think this PR is too ad-hoc for following reasons.

1) The following code should not raise `undefined method 'length' for 0:Integer (NoMethodError)` error.

```ruby
class C
  def initialize
    @v = [[0], 0]
  end
  
  def deconstruct
    @v.shift
  end
end

case C.new
in [1]
in []
end
```


2) The check for `#deconstruct` result type is omitted only for the top-level value.

```ruby
class C
  def initialize
    @v = [[0], 0]
  end
  
  def deconstruct
    @v.shift
  end
end

case [C.new]
in [[1]]
in [[]] # It checks result type. So it raises TypeError instead of NoMethodError, unlike the 1).
end
```

---

To improve performance, I plan to cache the results of `#deconstruct`.
There is no progress yet, so I'd be glad if you could implement it.