@palkan 2020-04-06
@k-tsj 

Regarding this example:

```
def deconstruct
  @v.shift
end
```

Ruby is to flexible by design, it's possible to find a counter-case for every optimization)

What if we can add more strictness to pattern matching? For example, state that we only call `#deconstruct` once, so the following patterns would behave similarly:

```
case v
in [1, 2]
in [3, 4]
end

dv = v.deconstruct
case dv
in [1, 2]
in [3, 4]
```

Actually, you proposed the idea of caching the `#deconstruct` result, which would introduce this strictness.

Adding a _contract_ to pattern matching would open a lot of opportunities for performance improvements.

> There is no progress yet, so I'd be glad if you could implement it.

I made an attempt to cache the `respond_to?` return value on the stack here: https://github.com/palkan/ruby/commit/ed51ffa9ef91b877acd09d57c3660c79cfd2abb1.
It could be modified to cache the `#deconstruct` result instead, I think.

The tests and specs pass but I'm not sure about the implementation: **is there a better way to cache an _internal_ (invisible from the Ruby code) value than reserving slots on the stack?**