@k-tsj 2020-04-12
> What if we can add more strictness to pattern matching?
>
> (snip)
>
> Actually, you proposed the idea of caching the #deconstruct result, which would introduce this strictness.

Yes, your understanding is correct.

> is there a better way to cache an internal (invisible from the Ruby code) value than reserving slots on the stack?

I think reserving slots on the stack is a best way.